backend/requirements.txt
fastapi
uvicorn
gunicorn
python-multipart
python-dotenv
pydantic
requests
beautifulsoup4
pandas
chromadb
sentence-transformers
pymupdf
unstructured[pdf]
python-docx
reportlab
serpapi
duckduckgo-search
pydantic-settings
langchain
langchain-groq
langchain-community
loguru
firebase-admin
psycopg2-binary
sqlalchemy
asyncpg
